requirements.txt
Django==2.0.1
django-cors-headers==2.1.0
numpy==1.14.0
python-magic==0.4.15
pytz==2017.3
scikit-learn==0.19.1
scipy==1.0.0
xmltodict==0.11.0
